Remove old Grid classes.
